Avoid allocations checking for global properties
When considering what to do about a property element, we avoid computing
values for global properties, since they can't be changed. That involves
querying whether the given property name is in the set of global
property names.

Prior to this commit, that code constructed a (new, used once)
`ReadOnlyConvertingCollection`, which involved allocating a
`Microsoft.Build.Collections.Contains` delegate. A trace observed that
52 MB of delegates were allocated in a NuGet scenario. None of that
allocation is necessary.

Fixes #2266 by changing `PropertyDictionary.ContainsKey()` to pass
through to its backing `RetrievableEntryHashSet` instead of going
through an intermediate `ReadOnlyConvertingCollection`, so no allocation
is required to check for the existence of a property.

There were no other callers of `RetrievableEntryHashSet.Keys`, so I
disabled it.