Fix unit-tests. MockEngine was caching the text of the log the first time it saw it, so subsequent calls to AssertLogContains were incorrect. Removing _upperLog.
This uncovered the two tests that were supposed to be failing, but are
passing due to the incorrect log being cached. Filed issue #2305 to fix
these tests.

Thread through the ITestOutputHelper so that the failing test can output
the full contents of the log in case of failure. ITestOutputHelper gets
injected into the constructor by xUnit.