Fallback culture translation for MSBuildTaskhost
When a CultureInfo is serialized on an MSBuild node (running on a newer .net framework) and sent to the MSBuildTaskHost (running on an older .net framework), the MSBuildTaskHost might not be able to deserialize the CultureInfo, as some culture codes are supported on newer .net framework versions but not on older ones.

Fallback to CultureInfo.CurrentCulture if CultureInfo cannot be instantiated on the culture code sent over from the msbuild node.