Remove namespace check of every XML element
Fixes #2373 by only checking the top-level namespace of a project, not
the namespace of every element.

This is arguably incorrect since XML namespace _could_ be per-element,
but a) that never worked before and b) you'd have to manually edit a
project in a fairly odd way to get that, so the performance gain here
seems to outweigh the risk.