Run InvalidOutputPath test only on Windows
This test was running everywhere, but when output was captured it
started failing because it was attempting to write a `\0` character
to the log stream, which confused the xunit xml logger and the
Jenkins parser.

Since it's rare to have "invalid" file names on non-Windows systems,
restrict this test to Windows instead.