[mono][tests] Use GetCommandLineArgs, also, to check for test runners
When running a test runner under mono, the process name is not helpful
to determine if we are running tests. Getting command line arguments,
OTOH, do help with this.

So, we bring back the old `GetProcessFromCommandLine()` method. And
also, add this as a parameter to the `ResetInstance_ForUnitTestsOnly(..)`
method.

This fixes
Microsoft.Build.Engine.UnitTests.BuildEnvironmentHelper_Tests.BuildEnvironmentDetectsRunningTests