[mono] install-mono-prefix.sh: Fix wrench build
`System.Reflection.Metadata.dll` in `bin/` needs to be same as
`bin/Roslyn/`. Earlier, this was installed to `bin/Roslyn/` from
msbuild, but now it is installed from mono.

But when bockbuild builds msbuild, it installs to a staging directory,
which doesn't have the roslyn bits from mono available, so we can't copy
the file. Instead we create a symlink. In a correct mono install, the
symlink would not be broken.