[mono] Fix loading CodeTaskFactory from v4/v12 task assembly
Mono's v4 and v12 assemblies as facade assemblies, with typeforwards to
the Core assembly. TypeLoader uses `Type.GetType` which doesn't return
forwarded types. So, loading task factory fails like:

    `<UsingTask TaskName="foo" TaskFactory="CodeTaskFactory" AssemblyFile="$(MSBuildToolsPath)\Microsoft.Build.Tasks.v4.0.dll">`

The reason it seems to work for loading tasks with `UsingTask`, is that
our default list of tasks references the `Core` assembly, so the issue
doesn't show up.

For the most common cases for task factory, we redirect to the Core
assembly, building on an existing workaround in the code. The better fix
would be to use Cecil or System.Reflection.Metadata to get the types.

We also fix the facades build to be before the tests. Otherwise, tests
depending on them fail.

Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=56633