Fix debian package generation
Previously, dotnet-deb-tool-consumer.csproj was copied to the out/artifacts folder before being restored.  This no longer works as it depends on repo properties, and there is a Directory.Build.props in teh out directory to stop projects inside it from getting the repo properties.  So this change restores it in-place instead.