Fix issues caused by adding RHEL6 support
There were two issues. One was that additional arguments that are not
recognized by the build.sh script were incorrectly passed as a single
argument in single-quotes to msbuild.
The second issue was that when LD_LIBRARY_PATH was not set, the
dotnet-install.sh script was failing due to its using of the "set -u"
option that prevents usage of unset variables.