[mono] Disables tests that are breaking xunit/osx (#27)
Some GenerateResource tests use `\0` as invalid file names on !windows.
And xunit tries to write the output of that build to xml (CDATA) without
filtering the NUL, and we get a ArgumentException.

Upstream worked around[1] this by disabling the tests. We are doing the
same here, till we have a build of xunit that fixes the issue.

Manifests on CI:
This issue causes the test build to fail as the exception is thrown
after all the tests have run and xunit is trying to write the xml
results file! So, wrench ends up with "failed" tests even though none of
the tests actually failed!

--
1. commit: 4603358e77ff5ebe46c33bbe21e7ae7d7ea7c642