Set document full path sooner to reduce allocations
Building a single project I saw 17,537 `ElementLocation` objects being created.  It turns out that code in `XmlElementWithLocation` and `XmlAttributeWithLocation` was making new objects because the document full path is not set in time.

This change sets the document full path sooner.  This reduced the number of ElementLocation objects being created from 17,537 to just 9,237 (47% less)