Avoid boxing KeyedObject
Fixes: #2561

We were boxing/allocating KeyedObject on every string lookup to RetrievableEntryHashSet, resulting in 0.4% -> 1.4% of allocations during a build.

To avoid that, changed RetrievableEntryHashSet to take IEqualityComparer<string> instead of IEqualityComparer<IKeyed>. Effectively, we were comparing strings underneath anyway - just moved to always comparing them. This removed all allocations on lookups to this collection.