Double Evaluation Fix (#2595)
* Get properties for ProjectReference in parallel

In order to avoid batching the
_GetProjectReferenceTargetFrameworkProperties target for each reference,
the ProjectReference protocol can be amended to return an item from
GetTargetFrameworkProperties instead of a semicolon-delimited list of
key-value pairs. This allows a single build request to be sent to the
engine, and allows resolving references in parallel on multiprocess
builds.

* Identify inner-build in consuming project
* Call GetTargetFrameworks target to identify target frameworks in
referenced projects. Since TargetFrameworks is not set, this will avoid
a 2nd evaluation when the project single targets.

* Adds dependency on NuGet GetReferenceNearestTargetFrameworkTask.

* Add NuGet ImportAfter targets to bootstrapped build
This is needed to get the import to get the automatic import of
NuGet.targets.