Avoid multiple GetLastWriteTime on same assembly
Fixes: #2440.

This makes sure that we only ever call GetLastWriteTime once during a single ResolveAssemblyReference session.

Note: I tried to maintain the existing behavior as much as possible except one change - previously, if a file on disk was found to be later than the process-wide cache, the process-wide entry would never get evicted (due to TryAdd).

This reduces CPU in GetFileState from 2% -> 0.7% in a mixed .NET Framework/.NET Standard solution.