Ship Nuget with the MsBuild.Engine.Corext package (#2656)
The common targets got a dependency on a new Nuget task, which ships
with VS15.5. Older MSBuild.Corext packages need to function with the
newest MSBuild.Engine.Corext package, so as a temporary workaround we're
including the nuget files. When VS15.5 ships, we should remove nuget, as
MSBuild.Corext 15.5 will bring the nuget files.