[mono] Fixes props and target imports in Xamarin.IoT.Sdk (#30)
MSBuildSdksPath was not being defined by certain VS Mac versions so the import was not working. This was being set as an environment variable but that is no longer the case. In more recent VS Mac versions, not yet published, the MSBuildSdksPath value is being set in the .exe.config for the MSBuild host. This was to fix a problem with building Xamarin.Forms 2.4 projects, which conditionally includes some MSBuild files if MSBuildSdksPath is defined, and also fixes problems with the Xamarin.IoT projects which were similar.