Prepare msbuild for orchestrated build Orchestrated build will pass parameters in a consistent manner from the top level.  This change prepare msbuild for this by dealing with two key issues:     * OSS signing - The sign type will be passed as 'public' from the root     * push to blob storage
To deal with signing, since the build tools version is so old (and only supports 'oss' as the sign type), I set the sign type to 'oss' internally if it comes in as public.  A new input parameter was used to obtain the value.  In the case of other build definitions not concerned with orchestrated build, the old behavior of SignType=oss will simply have the old behavior.  test/release values will operate as normal.  In addition, UseOpenSourceSign will be set appropriately if the sign type is public.  The value will be unaltered if the sign type is not 'public'.  This means existing behavior will be preserved.

I have also removed the oss satellite assembly target.  It's dead code, as the target it invokes (OpenSourceSign) is no longer imported.

For future refactors, we should probably unify on a single set of properties here (SignType), and remove all the various MSSignType/UseOpenSourceSign, etc. Unfortunately, becuase the build definitions used are currently not defined and branched with the msbuild source, making that change involves pushing through a change in multiple msbuild release branches.