Suppress assets messages for Clean target.
This commit ensures that messages from `project.assets.json` aren't logged for
the `Clean` target.  If a project is cleaned that has diagnostic messages
stored from a previous restore operation, the Clean target would previously
log the messages via the `ReportAssetsLogMessages` target.

Since a clean operation does not perform a restore, this is confusing to users.
The fix is to set the `EmitAssetsLogMessages` property to `false` for the
`Clean` target, but not the `Rebuild` target that depends on the `Clean`
target.

Fixes dotnet/cli#8027.