Breaking Change in FileMatcher: ab*ba pattern matches aba by mistake (#2814)
Performance optimization #2392 led to a breaking change in FileMatcher logic. Let's examine the following example with a single web.config file in working directory:

<MyFiles Include="**\*" Exclude="web.*.config">
</MyFiles>
Before the change @(MyFiles) included web.config because exclude pattern didn't match it. However after the change web.config was excluded from the item group. This is because implementation of FileMatcher.IsMatch() doesn't consider a situation when tail after wildcard (*) is longer than the remaining part of the input during "tail check". This is a pretty common scenario when several file extensions are used.

To be more specific FileMatcher.IsMatch("aba", "ab*ba") should return false instead of true.