[mono][osx] Fix test hangs
Some tests caused the test run to hang under mono/osx. This is because
of an `InternalErrorException` being thrown and msbuild trying to launch
and connect to a debugger. We do try to avoid the debugger code path
(`ConsiderDebuggerLaunch`), but it depends on
`BuildEnvironmentHelper.RunningTests` which incorrectly gets set to
`false`.

When running a test runner under mono, the process name is not helpful
to determine if we are running tests as that would be `mono`. Getting
command line arguments, OTOH, do help with this.

So, we use GetCommandLineArgs, also, to check for test runners. We bring
back the old `GetProcessFromCommandLine()` method. And also, add this as
a parameter to the `ResetInstance_ForUnitTestsOnly(..)` method.

Example test that triggered this:
    `Microsoft.Build.UnitTests.ErrorUtilities_Tests.VerifyThrowFalse`