In Expander.cs add a shortcut for some common string methods to bypass binding (#2833)
This avoids first-chance MissingMethodExceptions when evaluating most solutions, such as Roslyn, MonoDevelop and a few other ones I tested.

This is a tactical fix to get the exception out of the way for now. We can think of a proper fix to the binder later. This should also slightly improve evaluation performance - each binder call that throws costs ~1.5 ms on my machine, whereas this method brings it down to ~0.05 ms, so about 30x improvement.

I'm just including the functions I ran into during testing that are in the wild. This list is not exhaustive, because for rare cases we can still fallback to the old behavior, it's just going to be way less common now. I've verified that using this MSBuild in a certain test environment is able to load large solutions without seeing any of the first-chance exceptions.