A tab separated file is generated as the default format for the profiler report (#2837)
This change introduces a format change in the profiler report (/prof:<filename> option). If the specified filename is '.md', the existing markdown format is used. Otherwise, for any other extension, a tab separated file is generated. This allows users to open the profiler report in tools like Excel and manipulate the data directly.