No tfcompat check if ReferenceOutputAssembly=false
It is confusing to explicitly specify that you _don't_ want the output
of another project referenced in this project and then be told that the
output is incompatible.

This commit listens to the preexisting ProjectReference metadatum
ReferenceOutputAssembly and avoids the compatibility/best-match checks
on ProjectReferences that avoid the dependency.

Fixes #2661 (and dotnet/sdk#939).