Generate binding redirects for .NET 4.7.2+
In The Old Days, binding redirect generation was set to off-by-default.
That has caused a great deal of confusion because it's usually the right
thing to do but it's not obvious when you need to do it.

Unfortunately, we currently think we can't change the default for all
projects, because a user could have been avoiding binding redirects
intentionally, making creating them a breaking change.

Since no projects currently target .NET Framework 4.7.2 (since it's not
released), it is not a breaking change to turn on binding redirects by
default for this (and higher) versions--a user must explicitly opt into
targeting the framework and thus into this new behavior.

A user can still specify

```xml
<AutoGenerateBindingRedirects>false</AutoGenerateBindingRedirects>
```

to explicitly opt out of binding redirects for a given project.

Closes #2481, in partial support of

https://github.com/dotnet/designs/blob/2be39b7e1bc9c4df45b1d51ebb2d1abb0689e4e0/accepted/automatic-redirects-by-default/automatic-redirects-by-default.md