Remove System.Diagnostics.Process from CoreXT pkg (#2900)
This doesn't appear to be used and is not copied by the build. No idea
how this worked before, sync to the last successful build also does not
produce this file. This is failing our official build.

MSBuild only references System.Diagnostics.Process for netstandard, not
for net46 so appears to have been added by mistake.