Fix preprocessor when imported documents specify Sdk
Implicit imports for imported projects are shown in the preprocessor at the top and bottom of the top level project rather than the imported project.

To fix this, I've created a mapping of which documents contain the implicit imports.  Before processing a document, the imports are then added to that document rather than the top level one.