Remove netstandard1.3 reference
Without building 3 times we can't really support netstandard1.3 as we start to expose types that are public in the full framework version but not available in netstandard1.3 (for example public binary serialization fields).

For PortableTask to be "portable" to older versions it needs to reference our 15.5.180 NuGet package which contains netstandard1.3 support.