Stop using static constructor to wire up AssemblyResolve event handler in NuGetSdkResolver (#2925)
My original intent was to wire up the AssemblyResolve event as soon as possible.  Later I moved all NuGet type references to private classes instead which means that NuGet assemblies are not loaded until these private classes are accessed.  The private classes are not accessed unless the user specified a version.

So wiring up AssemblyResolve in a static constructor means that in Visual Studio all AssemblyResolve events are sent through the NuGet SDK resolver which is no ideal.  Instead, it can be wired up prior to resolving an SDK and then removed.

I also added a unit test to ensure that NuGet assemblies are not loaded when no version is specified.