Fix node task crash in SDK resolution
In a multi-proc build when an SDK could not be found, a NullReferenceException would occur which would crash the thread on the main node responsible for responding to SDK resolution requests.  The out-of-proc node would then hang waiting for a response and MSBuild's main thread would be waiting for the out-of-proc node to finish.

The code is more resilient now by catching all exceptions and sending the response in a finally.  Any errors that occur in the request handler thread are sent back to the out-of-proc node to be logged.