Fix node task crash in SDK resolution which hangs MSBuild.exe (#2964)
* Fix node task crash in SDK resolution

In a multi-proc build when an SDK could not be found, a NullReferenceException would occur which would crash the thread on the main node responsible for responding to SDK resolution requests.  The out-of-proc node would then hang waiting for a response and MSBuild's main thread would be waiting for the out-of-proc node to finish.

The code is more resilient now by catching all exceptions and sending the response in a finally. 

* Log on the main node since it has the logging context and make it a fatal build error since its unexpected

* Call LogFatalBuildError so I don't need a new message

LogFatalBuildError includes exception.ToString()

* Fix bad version check