(Re)set ComVisible(false)
Prior to the move to GitHub, MSBuild types were by default not
ComVisible (with selective opt-in visibility). This was established by a
common VS repo setting, and not obvious in the MSBuild source.

Having everything be default visibility caused problems when updating
COM references to MSBuild within the VS repo--types started conflicting,
because they shouldn't have been exposed.