Create COM typelibs; include them in ExternalAPIs
Several parts of VS that are implemented in native code use COM
to call into MSBuild, and thus depend on our .tlbs. In the VS repo,
we were building these using common VS infrastructure by having

    <CreateTlb>true</CreateTlb>

in our projects, but that did nothing once we pulled out of VS.

I'm adding it back to the new projects and adding a target to pay
attention to it. The target requires that `tlbexp.exe` be available
in PATH.