Add test that RemoveDuplicates preserves order
The RemoveDuplicates task can be made more useful to consumers if its
contract is strengthened to say that the Filtered output is guaranteed
to be in the same order as its Inputs. It's current implementation has
this behavior, so additional unit tests have been added to ensure this
does not regress.

Related: https://github.com/MicrosoftDocs/visualstudio-docs/issues/622