Wait on active target in another RequestBuilder
If a target is marked as actively building by one RequestBuilder but
gets to the top of the target stack in another RequestBuilder, we would
throw an exception that was swallowed, failing the project silently.

The fix is to check if the target is actively building before attempting
to start it. If it is, wait on its results before proceeding.