Spawn nodes without handle inheritance on Win/Core
On Windows, processes by default inherit standard stream handles from
their parent process. This can cause a process that is waiting on
termination of a given process to wait on its children as well while
waiting for the streams to close.

We want our worker nodes (when launched with `/nodeReuse:true`) to
entirely disconnect from their parent and live past its lifetime. On
Windows, that means we need to avoid inheriting handles.

We had code to p/invoke `CreateProcess` instead of using
`Process.Start`, but it was #ifdef'ed to full framework only. Changing
that to be a runtime check for `IsWindows`.

The stdout/stdin/stderr handles are manually disconnected, but handles
for IPC must be inherited or the child node can't connect to its parent,
so pass `bInheritHandles=true` for the anonymous-pipe communication case
(in netcoreapp2.0).

Fixes #3091.