Speed up incremental package asset resolution (#2020)
Whenever project.assets.json changes (or more rarely, a setting influencing
ResolvePackageAssets changes), write out a binary cache file for the next
incremental build. The file has only the items that used by the build and
they are written in a format that can be deserialized to MSBuild ITaskItem[]
with minimal time and allocation.

On top of that, we merge ReportAssetsLogMessages into ResolvePackageAssets, 
which allows for project.assets.json to not be read at all if it has not
changed.

However, design time build still uses old ReportAssetsLogMessages like old
ResolvePackageDepdencies. Fixing the design time perf will be done separately
in coordination with project system.

In order to ensure consistent behavior between incremental builds and full
builds, and to ensure maximal test coverage of cache reading and writing, a 
full build will write out the cache and then read from it as an incremental
build would.