Use subfolder under TEMP for test transients
This will have a shorter path on *nix, helping avoid the 108 character
pipe path limit there.

Prior to this change, `Directory.Build.*` was read from the test output
directory, because TEMP was a subfolder of that. This changed that, so
create empty versions of each file to stop recursing upward (and
possibly finding the repo-root one, causing incredible headache).

Co-authored-by: Mihai Codoban <micodoba@microsoft.com>