Fix XmlDocument reader to allow opening files in read-only mode
XmlDocumentWithLocation class has the ability to openXML documents in the read-only mode (comments and whitespace is replaced with an empty string). Currently this functionaly is not available, because the XML file location is not passed to the XML reader constructor and thus not available for the XmlDocumentWithLocation for further verification.
Additionally, the IsReadOnly property is now overriden, to correctly inform callers about the read mode of the document.
Added unit tests that cover these scenarios.