Fix case sensitivity when mapping project paths to project infos.
When a solution contains a path to a project that differs by case for a project
reference identity, the build fails because the project info can't be found.

The fix is to make the internal map case insensitive, allowing the case
difference in a project's path in the solution file.

Fixes dotnet/cli#8048.