Move DefineConstants where they'll actually apply
We were setting it before Sdk.props / Directory.Build.props, and
repo toolset was overwriting it.

This caused #if EXTENSIONS to never be taken and a task loading error
to occur when NETBuildExtensionsError was used.

This is the minimal fix for 2.1.2xx / 15.7. I hope to refactor things
in 2.1.3xx so that our project files are more standard and this mistake
is harder to make.