Assume node reuse in ShutdownAllNodes
ShutdownAllNodes() has two functions: it can close stale background
nodes left over from previous uses, or, if you're manually managing a
build, it can indicate that you'd like to close all the nodes that you
started up, regardless of node-reuse status.

Changes for #346 introduced a regression in the former case, in the
situation where BuildParameters is null. In that case, we can safely
assume that we're being asked to shut down all reachable nodes that are
idle, and they would only be idle if they were launched with node reuse.
So assume nr:true.

Fixes #1467