Only write the generated app.config if needed (#3009)
* Only write the generated app.config if needed

We compare what we're going to write to what is already there if
present. If the contents are already there, then don't bother.

Note: This is rather naive, as any difference will cause the file to be
written. For example if, for some reason, the order of the binding
redirects changes.

This fixes #3008

* GenerateBindingRedirectsTests code cleanup

* Add unit test