Disconnect stdout/stdin/stderr from worker nodes
`Process.WaitForExit()` waits for the process's I/O streams to close
before returning. If that process spawned its own grandchild process,
the grandchild by default shares its parents' streams. This caused long
hangs when running MSBuild with node reuse:

* From some launching process
* Spawn a top-level MSBuild process
* Wait on its return
* It starts some new worker processes
* Build finishes
* Top-level MSBuild exits
* Launching process continues to wait on stream close, appearing to hang
* MSBuild worker nodes hit idle timeout after 15 minutes
* Launching process resumes

This was solved on Windows by using win32 `CreateProcess` directly and
passing invalid handles to the worker node's streams, which broke the
chain. But it still happened on non-Windows OSes.

This commit fixes that by redirecting the streams of the worker nodes
but not doing anything with them. It's essentially the same as
https://github.com/dotnet/roslyn/pull/23805.